requirements.txt
Flask
fpdf
pandas
pdfplumber
googlesearch-python

